requirements.txt
playwright==1.42.0
google-genai==0.3.0
python-dotenv==1.0.1
